# stdenv.mkDerivation: warn when overriding version without also overriding src