@winterqt 2025-05-12
Will fix, my inlining wasn't 100% correct I guess.