@winterqt 2025-05-12
> > this doesn't have any effect due to how most of our FODs are named.
> 
> Are there any packages that provide a hash arg that technically wouldn't require touching `src`?

Yes, see the most recent push. (I was in the process of fixing this case when you commented.)