@winterqt 2025-05-12
No idea why the attrpaths eval caught everything but the OBS Studio warning, will fix.