@RossComputerGuy 2025-05-12
> > sounds like it's something in a similar realm to #317663.
> 
> Yeah, good point -- though this is obviously not _the_ solution, just a footgun remover. (Though I really do want to stop people from using `finalAttrs.version` like that, maybe some day it'll come to fruition.)

Yeah, just wanted to link it for visibility. Imo, this is an incremental step towards it.