@winterqt 2025-05-15
I have no idea why the [perf comparison](https://github.com/NixOS/nixpkgs/actions/runs/15051319861/attempts/1#summary-42306868879) doesn’t show such a thing, or why some %s are uh… negative, when they very obviously *increase*?