@winterqt 2025-05-15
I ran some benchmarks:

### Overlay without `version`

```
Benchmark 1: nix-instantiate --expr --eval "((import ./nixpkgs { }).hello.overrideAttrs (_: { a = 1; })).outPath"
  Time (mean ± σ):     256.5 ms ±   0.7 ms    [User: 206.8 ms, System: 44.2 ms]
  Range (min … max):   255.4 ms … 257.8 ms    11 runs

Benchmark 2: nix-instantiate --expr --eval "((import ./nixpkgs-before { }).hello.overrideAttrs (_: { a = 1; })).outPath"
  Time (mean ± σ):     256.8 ms ±   1.0 ms    [User: 206.8 ms, System: 44.2 ms]
  Range (min … max):   255.7 ms … 259.0 ms    11 runs

Summary
  nix-instantiate --expr --eval "((import ./nixpkgs { }).hello.overrideAttrs (_: { a = 1; })).outPath" ran
    1.00 ± 0.00 times faster than nix-instantiate --expr --eval "((import ./nixpkgs-before { }).hello.overrideAttrs (_: { a = 1; })).outPath"
```

### Overlay with `version`

```
Benchmark 1: nix-instantiate --expr --eval "((import ./nixpkgs { }).hello.overrideAttrs (_: { version = \"meow\"; })).outPath"
  Time (mean ± σ):     257.3 ms ±   0.5 ms    [User: 206.9 ms, System: 44.8 ms]
  Range (min … max):   256.2 ms … 257.9 ms    11 runs

Benchmark 2: nix-instantiate --expr --eval "((import ./nixpkgs-before { }).hello.overrideAttrs (_: { version = \"meow\"; })).outPath"
  Time (mean ± σ):     257.6 ms ±   1.1 ms    [User: 207.1 ms, System: 44.6 ms]
  Range (min … max):   256.5 ms … 260.3 ms    11 runs

Summary
  nix-instantiate --expr --eval "((import ./nixpkgs { }).hello.overrideAttrs (_: { version = \"meow\"; })).outPath" ran
    1.00 ± 0.00 times faster than nix-instantiate --expr --eval "((import ./nixpkgs-before { }).hello.overrideAttrs (_: { version = \"meow\"; })).outPath"
```

### Overlay with both `version` and `src`

```
Benchmark 1: nix-instantiate --expr --eval "((import ./nixpkgs { }).hello.overrideAttrs (_: { version = \"meow\"; src = \"\"; })).outPath"
  Time (mean ± σ):     253.7 ms ±   0.6 ms    [User: 204.0 ms, System: 44.3 ms]
  Range (min … max):   252.8 ms … 254.6 ms    11 runs

Benchmark 2: nix-instantiate --expr --eval "((import ./nixpkgs-before { }).hello.overrideAttrs (_: { version = \"meow\"; src = \"\"; })).outPath"
  Time (mean ± σ):     254.5 ms ±   0.6 ms    [User: 204.5 ms, System: 44.7 ms]
  Range (min … max):   253.6 ms … 256.0 ms    11 runs

Summary
  nix-instantiate --expr --eval "((import ./nixpkgs { }).hello.overrideAttrs (_: { version = \"meow\"; src = \"\"; })).outPath" ran
    1.00 ± 0.00 times faster than nix-instantiate --expr --eval "((import ./nixpkgs-before { }).hello.overrideAttrs (_: { version = \"meow\"; src = \"\"; })).outPath"
```