@ElvishJerricco 2025-05-15
Can you try something with more transitive dependencies? i.e. More `mkDerivation` calls to potentially slow down? e.g. Firefox or systemd or something.