@winterqt 2025-05-15
### Without `version`

```
Benchmark 1: nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs { }).systemd.overrideAttrs (_: { a = 1; })).outPath"
  Time (mean ± σ):     364.6 ms ±   1.8 ms    [User: 306.8 ms, System: 51.4 ms]
  Range (min … max):   363.3 ms … 369.4 ms    10 runs

Benchmark 2: nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs-before { }).systemd.overrideAttrs (_: { a = 1; })).outPath"
  Time (mean ± σ):     368.2 ms ±   4.6 ms    [User: 309.3 ms, System: 52.0 ms]
  Range (min … max):   364.5 ms … 379.1 ms    10 runs

Summary
  nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs { }).systemd.overrideAttrs (_: { a = 1; })).outPath" ran
    1.01 ± 0.01 times faster than nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs-before { }).systemd.overrideAttrs (_: { a = 1; })).outPath"
```

### With just `version`

```
Benchmark 1: nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs { }).systemd.overrideAttrs (_: { version = \"meow\"; })).outPath"
  Time (mean ± σ):     364.6 ms ±   2.2 ms    [User: 307.0 ms, System: 51.7 ms]
  Range (min … max):   362.0 ms … 368.8 ms    10 runs

Benchmark 2: nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs-before { }).systemd.overrideAttrs (_: { version = \"meow\"; })).outPath"
  Time (mean ± σ):     365.1 ms ±   2.4 ms    [User: 307.1 ms, System: 50.9 ms]
  Range (min … max):   363.0 ms … 369.4 ms    10 runs

Summary
  nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs { }).systemd.overrideAttrs (_: { version = \"meow\"; })).outPath" ran
    1.00 ± 0.01 times faster than nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs-before { }).systemd.overrideAttrs (_: { version = \"meow\"; })).outPath"
```

### With `version` and `src`

```
Benchmark 1: nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs { }).systemd.overrideAttrs (_: { version = \"meow\"; src = \"\"; })).outPath"
  Time (mean ± σ):     364.2 ms ±   2.4 ms    [User: 306.7 ms, System: 51.5 ms]
  Range (min … max):   362.5 ms … 370.4 ms    10 runs

Benchmark 2: nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs-before { }).systemd.overrideAttrs (_: { version = \"meow\"; src = \"\"; })).outPath"
  Time (mean ± σ):     364.8 ms ±   1.9 ms    [User: 306.7 ms, System: 52.5 ms]
  Range (min … max):   361.9 ms … 368.3 ms    10 runs

Summary
  nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs { }).systemd.overrideAttrs (_: { version = \"meow\"; src = \"\"; })).outPath" ran
    1.00 ± 0.01 times faster than nix-instantiate --option eval-system x86_64-linux --expr --eval "((import ./nixpkgs-before { }).systemd.overrideAttrs (_: { version = \"meow\"; src = \"\"; })).outPath"
```